% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aCount.R
\name{aCount}
\alias{aCount}
\title{Function used when converting allele formats for Bayesian Genomic Clines admixed file.}
\usage{
aCount(x)
}
\arguments{
\item{x}{flattend allele dataframe}
}
\description{
BGC admixed allele conversion
}

