% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlleleFreq.R
\name{AlleleFreq}
\alias{AlleleFreq}
\title{Calculate the major allele}
\usage{
AlleleFreq(x)
}
\arguments{
\item{x}{Vector of characters representing alleles}
}
\description{
sub-function used by genepop_allelefreq
}

