% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlleleFreqLoci.R
\name{AlleleFreqLoci}
\alias{AlleleFreqLoci}
\title{Calculate allele frequency for a given allele}
\usage{
AlleleFreqLoci(x, major)
}
\arguments{
\item{x}{Vector of characters representing alleles}

\item{major}{is the major allele for a given loci}
}
\description{
sub-function used by genepop_allelefreq
}

