% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_string.R
\name{common_string}
\alias{common_string}
\title{Identify common string sequence}
\usage{
common_string(Vec)
}
\arguments{
\item{Vec}{A character vector.}
}
\description{
A function which returns the largest common sequence among a vector of strings.
}

