% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_check.R
\name{filename_check}
\alias{filename_check}
\title{filename_check}
\usage{
filename_check(X)
}
\arguments{
\item{X}{what to check}
}
\description{
Gets the file name that is specified by GenePop variable. This function is used by various functions of genepopedit and is applied to the Genepop variable.
}

