% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{frequencies}
\alias{frequencies}
\title{Calculate allele frequencies - sub-function of genepop_allelefreq()}
\usage{
frequencies(x)
}
\arguments{
\item{x}{Vector of characters representing alleles}
}
\description{
sub-function used by genepop_allelefreq
}
