% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_clean.R
\name{genepop_clean}
\alias{genepop_clean}
\title{Fixes common formatting problems using PGDspider.}
\usage{
genepop_clean(genepop, where.pgdspider, allocate.pgd.ram = 1,
  sample_delim = TRUE, population_suffix = FALSE,
  pop_sample_delim = FALSE, path)
}
\arguments{
\item{genepop}{A file path to the genepop format file you wish to create your panel from}

\item{where.pgdspider}{A file path to the PGDspider installation folder.}

\item{allocate.pgd.ram}{An integer value in GB to specify the maximum amount of RAM to allocate to PGDspider. The default is 1 GB, which should be sufficient for most analyses.}

\item{sample_delim}{A logical variable (default = TRUE) that defines whether the function needs to fix the sampleID loci delimiter using PGDspider. Here all files require (space,space scpace) to delimit the sampleIDs from the loci}

\item{population_suffix}{A logical variable (default = FALSE) that defines whether the function needs to add a population specific suffix to each population. This is required for some genepopedit functionality. Between each "pop" label a Pop(n)_ suffix will be added to each sampleID of genepop.}

\item{pop_sample_delim}{A logical variable (default = FALSE) that defines whether the function needs to add separation ("_") between population name and sample number. Note if population suffix = TRUE this option is redundant and not executed.}

\item{path}{the filepath and filename of output.#' @rdname genepop_toploci}
}
\description{
Use PGDspider to impliment the proper delimiters in your genepop file.
}
