% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_colony.R
\name{genepop_colony}
\alias{genepop_colony}
\title{Convert Genepop to Colony format.}
\usage{
genepop_colony(GenePop, where.PLINK, where.PGDspider, denote.missing = "000",
  allocate.PGD.RAM = 1)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab separation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will the standard genepop format with a the first n+1 rows corresponding the the n loci names,
or a single comma deliminated row of loci names followed by the locus data. Populations are
separated by "Pop". Each individual ID is linked to the locus data by " ,  " (space,space space) and is read in as
as a single row (character).}

\item{where.PLINK}{A file path to the PLINK installation folder.}

\item{where.PGDspider}{A file path to the PGDspider installation folder.}

\item{denote.missing}{The value that denotes missing data in your input file}

\item{allocate.PGD.RAM}{An integer value in GB to specify the maximum amount of RAM to allocate to PGDspider. The default is 1 GB, which should be sufficient for most analyses.}
}
\description{
Function to convert Genepop to Colony format.
}

