% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_detective.R
\name{genepop_detective}
\alias{genepop_detective}
\title{Explore Genepop data structure.}
\usage{
genepop_detective(GenePop, variable = "Pops")
}
\arguments{
\item{GenePop}{the genepop file to be manipulated. This will the standard
genepop format with a the first n+1 rows corresponding the the n loci names
followed by the locus data. Populations are seperated by "Pop".
Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character)
e.g.
Stacks Ver 1.0
1
2
3
Pop
Pop01_01  , 120120 110110 110110
Pop01_02  , 100100 110110 110110
Pop
Pop02_01  , 120120 110110 110110
...}

\item{variable}{data to be returned
Four options \code{default = "Pops"}
"Pops" = vector of population names
"PopNum" = dataframe of population names and counts
"Inds" = vector of sample IDs
"Loci" = vector of Loci}
}
\description{
Funtion returns Genepop file meta-data.
}

