% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_filename.R
\name{genepop_filename}
\alias{genepop_filename}
\title{genepop_filename}
\usage{
genepop_filename(X)
}
\arguments{
\item{X}{what to check}
}
\description{
Gets the file name that is specified by GenePop variable
}

