% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_filter_maf.R
\name{genepop_filter_maf}
\alias{genepop_filter_maf}
\title{Filters genepop file for MAF}
\usage{
genepop_filter_maf(genepop, where.plink, where.PGDspider, maf = 0.05, path)
}
\arguments{
\item{genepop}{A file path to the genepop format file you wish to filter.}

\item{where.plink}{A file path to the PLINK installation folder.}

\item{maf}{Minor allele frequency cutoff (default = 0.05)}

\item{where.pgdspider}{A file path to the PGDspider installation folder.}
}
\description{
Converts genepop file to PED and filters for MAF using plink and returns genepop file with only SNPs that pass filter
}

