% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_flatten.R
\name{genepop_flatten}
\alias{genepop_flatten}
\title{Flatten Genepop to dataframe}
\usage{
genepop_flatten(GenePop)
}
\arguments{
\item{GenePop}{the genepop file to be manipulated. This will the standard
genepop format with a the first n+1 rows corresponding the the n loci names
followed by the locus data. Populations are seperated by "Pop".
Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character)
e.g.
Stacks Ver 1.0
1
2
3
Pop
Pop01_01  , 120120 110110 110110
Pop01_02  , 100100 110110 110110
Pop
Pop02_01  , 120120 110110 110110
...}
}
\description{
Convert and return Genepop as flattend dataframe.
}

