% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_fstat.R
\name{genepop_fstat}
\alias{genepop_fstat}
\title{Convert Genepop to FSTAT format.}
\usage{
genepop_fstat(GenePop, path = NULL, addworkspace = FALSE)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single commma deliminated row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked to the locus data by "   , " and is read in as
as a single row (character).}

\item{path}{the filepath and filename of output.}

\item{addworkspace}{logical statement defining whether the converted data
should be saved as a file specified in the path (default) arguement or whether it should be returned to the workspace
if returned to the workspace the object will be called "Output_fstat".}
}
\description{
Function to convert Genepop to FSTAT
}

