% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_gsisim.R
\name{genepop_GSIsim}
\alias{genepop_GSIsim}
\title{Convert Genepop to GSI sim format.}
\usage{
genepop_GSIsim(genepop, path)
}
\arguments{
\item{genepop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab separation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single comma delimited row of loci names followed by the locus data. Populations are
separated by "Pop". Each individual ID is linked to the locus data by " ,  " (space, space space) and is read in as
as a single row (character).}

\item{path}{the filepath and filename of output.}
}
\description{
Covert from GENEPOP to format required by gsi_sim. Note that output has SampleIDs formated as Population_Population_ID instead of conventioanl Population_ID to fit commong sampleID naming approach of genepopedit into the convention of gsi_sim.
}
