% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_hzar.R
\name{genepop_hzar}
\alias{genepop_hzar}
\title{Convert Genepop to HZAR format.}
\usage{
genepop_hzar(genepop, distances, path)
}
\arguments{
\item{genepop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab separation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single comma delimited row of loci names followed by the locus data. Populations are
separated by "Pop". Each individual ID is linked to the locus data by " ,  " (space,space space) and is read in as
as a single row (character).}

\item{distances}{A dataframe or path to a text file with your distances between populations. Should contain 2 columns -
Populations and Distances.There should be the same number of populations as in the Genepop file.}

\item{path}{the filepath and filename of output.}
}
\description{
Function to convert Genepop to the R package HZAR.
}
