% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_ID.R
\name{genepop_ID}
\alias{genepop_ID}
\title{Add seperation ("_") between population name and sample number}
\usage{
genepop_ID(GenePop, path)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single commma delimited row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked to the locus data by "   , " (space,space space) and is read in as
as a single row (character).}

\item{path}{the filepath and filename of output.}
}
\description{
Function to add seperation ("_") between population name and sample number
}

