% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_toploci.R
\name{genepop_toploci}
\alias{genepop_toploci}
\title{Creates a panel of the top n unlinked loci, and exports the list of loci}
\usage{
genepop_toploci(GenePop, LDpop = "Both", panel.size = NULL, where.PLINK,
  where.PGDspider)
}
\arguments{
\item{GenePop}{A file path to the GENEPOP format file you wish to create your panel from}

\item{LDpop}{A string which denotes which of the two populations you wish to calculate linkage disequilibrium from. The options are "Pop1" or "Pop2", or "Both" (default) if the LD is to be calculated based on both populations. The order of population is based on the order by which they appear in the genepopfile. refer to genepop_detetive() to check which population is considered 'Pop1' and 'Pop2'.}

\item{panel.size}{An integer number of loci to include in the panel. If not specified all loci will be returned,}

\item{where.PLINK}{A file path to the PLINK installation folder.}

\item{where.PGDspider}{A file path to the PGDspider installation folder.}
}
\description{
Extract the genotypes of individuals at the top n (by Fst) unlinked loci. The default threshold of r2>0.2 is employed for assigning 'linked' loci (default for plink).
}

