% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_treemix.R
\name{genepop_treemix}
\alias{genepop_treemix}
\title{Convert a Genepop file for input in the program TREEMIX}
\usage{
genepop_treemix(GenePop, where.PGDspider, where.PLINK, binaryout,
  allocate.PGD.RAM)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab separation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard Genepop format with the first n+1 rows corresponding to the n loci names,
or a single comma delimited row of loci names followed by the locus data. Populations are
separated by "Pop". Each individual ID is linked to the locus data by " ,  " (space,space space) and is read in as
as a single row (character).}

\item{where.PGDspider}{A file path to the PGDspider installation folder.}

\item{where.PLINK}{A file path to the PLINK installation folder.}

\item{binaryout}{A name that you want to call your binar PED file created from PLINK}

\item{allocate.PGD.RAM}{An integer value in GB to specify the maximum amount of RAM to allocate to PGDspider. The default is 1 GB, which should be sufficient for most analyses.}
}
\description{
Convert Genepop to within-clusters binary PED file for TREEMIX
}

