% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_unflatten.R
\name{genepop_unflatten}
\alias{genepop_unflatten}
\title{Convert to Genepop format from a flattened dataframe.}
\usage{
genepop_unflatten(df, path)
}
\arguments{
\item{df}{dataframe with the first column holding sampleIDs (e.g. BON_01) and the remaining columns holding loci. Column names of loci will be used as loci names in the genepop output.
df must be an object in the workspace.}

\item{path}{the filepath and filename of output.}
}
\description{
Function returns Genepop file meta-data.
}

