% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_vals.R
\name{insert_vals}
\alias{insert_vals}
\title{Insert values at location}
\usage{
insert_vals(Vec, breaks, newVal)
}
\arguments{
\item{breaks}{vector of positions where values will be inserted.
Here the output will have length(vec)+length(breaks) rows with the values
specified in 'newVal' positioned at the 'breaks' positions}

\item{newVal}{values to be inserted}

\item{vec}{A data vector where values will be inserted.}
}
\description{
Function to insert values at specified positons within a dataframe and-or vector
}

