% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majorminor.R
\name{majorminor}
\alias{majorminor}
\title{Function to convert allele in character format to bgc zygosity format.}
\usage{
majorminor(Vec, allele_length = 6)
}
\arguments{
\item{Vec}{Vector of characters representing alleles.}

\item{allele_length}{The number of characters in an allele (default: 6 e.g. 130120).}
}
\description{
sub-function used by genepop_bgc
}
