% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microhp_genepop.R
\name{micro_genepop}
\alias{micro_genepop}
\title{Convert to genepop from micro haplo plot file format.}
\usage{
micro_genepop(microhp, path)
}
\arguments{
\item{microhp}{This is a file path to the micro-haplo-plot file. See update for link to this file format.}

\item{path}{the filepath and filename of output.}
}
\description{
Function makes a genepop file.
}

