% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgdSpideR.R
\name{PGDspideR}
\alias{PGDspideR}
\title{Execute PGDspider data format conversions in R}
\usage{
PGDspideR(input, input_format, output, output_format, spid, where.pgdspider)
}
\arguments{
\item{input}{the full path to the input file.}

\item{input_format}{what are you converting from (e.g. GENEPOP or PED) these must match the dropdown menus in pgdSpider.}

\item{output}{the full file path to the desired output. (e.g. GENEPOP or PED) these must match the dropdown menus in pgdSpider.}

\item{output_format}{what are you converting to (e.g. GENEPOP or PED). This must match the dropdown menus in pgdSpider.}

\item{spid}{the fill file path to the spid file generated in PGDspider.}

\item{where.pgdspider}{A file path to the PGDspider installation folder.}
}
\description{
Function to between file types in R.
}
