% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genepop_individuals.R
\name{subset_genepop_indiv}
\alias{subset_genepop_indiv}
\title{Genepop remove or keep specific sample IDs}
\usage{
subset_genepop_indiv(GenePop, indiv = NULL, keep = FALSE, path)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will the standard genepop format with a the first n+1 rows corresponding the the n loci names,
or a single commma deliminated row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character).}

\item{keep}{logical whether to delete sample IDs specified by indiv (default: TRUE) or delete all other IDs.}

\item{path}{the filepath and filename of output}

\item{vector}{sample IDs of interest.
These can be either the order by which they occur or the exact name of the loci
indiv <- \code{c("Pop01_01","Pop03_15","Pop16_02")} would individuals with these sample names.}
}
\description{
Function for the manipulation of genopop format SNP datasets
}

