% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genepop_individual.R
\name{subset_genepop_individual}
\alias{subset_genepop_individual}
\title{Genepop remove or keep specific sample IDs}
\usage{
subset_genepop_individual(GenePop, indiv = NULL, keep = FALSE, path)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single commma delimited row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked to the locus data by " ,  " (space,space space) and is read in as
as a single row (character).}

\item{indiv}{vector sample IDs of interest.
These can be either the order by which they occur or the exact name of the loci
indiv <- \code{c("Pop01_01","Pop03_15","Pop16_02")} would individuals with these sample names.}

\item{keep}{logical whether to delete sample IDs specified by indiv (default: TRUE) or delete all other IDs.}

\item{path}{the filepath and filename of output.}
}
\description{
Function for the manipulation of genopop format SNP datasets
}

