% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genepop_rename.R
\name{subset_genepop_rename}
\alias{subset_genepop_rename}
\title{Genepop subset and rename populations}
\usage{
subset_genepop_rename(GenePop, dirname, nameframe)
}
\arguments{
\item{GenePop}{the genepop file to be manipulated. This will the standard
genepop format with a the first n+1 rows corresponding the the n loci names
followed by the locus data. Populations are seperated by "Pop".
Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character)
e.g.
Stacks Ver 1.0
1
2
3
Pop
Pop01_01  , 120120 110110 110110
Pop01_02  , 100100 110110 110110
Pop
Pop02_01  , 120120 110110 110110
...}

\item{dirname}{path to the output .txt file (e.g."c:/Users/Yourname/Documents/output.txt").}

\item{nameframe}{a dataframe or path to a csv. This must be specified in this function
this dataframe contains two columns. Column 1 corresponds to the population names. These names
should match the individual IDs (e.g. BON01  , 110110 120120 -- would be 'BON'). The next column
has the new names that you want to change or rename. If you don't want to change the name then just repeat
from column one in that row. **Note if this is a reference to a path then there should be column headers
though the headers do not need to match this example. If the input is a dataframe object from the
workspace it must be a data.frame object and therefore will have headers.
This function assumes that each population will have a unique name. Names in this case are
comprised of alpha characters and not numbers.
(e.g. Pop01_01 and Pop02_01 would each be considered 'Pop' for the population name)
e.g.
Opop   Rename
BON    BON
BRA    BON
EDN    EDN
CRA    CRA
MAL    BON
TRY    CRA}
}
\description{
Function for the manipulation of genopop format SNP datasets and renaming of populations
}

