% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlleleFreq.R
\name{AlleleFreq}
\alias{AlleleFreq}
\title{Calculate allele frequencies}
\usage{
AlleleFreq(x, Freq = "Major")
}
\arguments{
\item{x}{Vector of characters representing alleles}

\item{Freq}{vector defining which allele frequency is returned (Major, Minor or Missing)}
}
\description{
sub-function used by genepop_allelefreq
}

