% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_allelefreq.R
\name{genepop_allelefreq}
\alias{genepop_allelefreq}
\title{Explore population specific allele frequencies.}
\usage{
genepop_allelefreq(GenePop, popgroup = NULL)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single commma delimited row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked to the locus data by " ,  " (space space,space) and is read in as
as a single row (character).}

\item{popgroup}{population grouping using the "Pop" deliminiter (Default: NULL) or a dataframe or path to a csv. The grouping dataframe should have two columns, the first corresponding to the population name and the second to an aggregation vector of common groups. Each population can only be assigned to one group.}
}
\description{
Function returns population derived allele frequencies.
}

