% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_detective.R
\name{genepop_detective}
\alias{genepop_detective}
\title{Explore Genepop data structure.}
\usage{
genepop_detective(GenePop, variable = "Pops")
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will the standard genepop format with a the first n+1 rows corresponding the the n loci names,
or a single commma deliminated row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character).}

\item{variable}{data to be returned
Four options \code{default = "Pops"}
"Pops" = vector of population names.
"PopNum" = dataframe of population names and counts.
"Inds" = vector of sample IDs.
"Loci" = vector of Loci.
"Allele" = vector of allele values.}
}
\description{
Funtion returns Genepop file meta-data.
}

