% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_allelefreq.R, R/genepop_detective.R
\name{genepop_allelefreq}
\alias{genepop_allelefreq}
\alias{genepop_detective}
\title{Explore population specific allele frequencies.}
\usage{
genepop_allelefreq(GenePop, popgroup = NULL)

genepop_detective(GenePop, variable = "Pops")
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single commma delimited row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked to the locus data by " ,  " (space space,space) and is read in as
as a single row (character).}

\item{popgroup}{if specified (Default: NULL) popgroup is a dataframe or path to a csv.
This dataframe contains two columns. Column 1 corresponds to the population names. These names
should match the individual IDs (e.g. BON_01 ,  110110 would be 'BON'). The next column
has the group. If groupings are the same as populations then leave as NULL (Default). Populations will be defined in this function using the population grouping format in the GENEPOP file.}

\item{variable}{data to be returned
Four options \code{default = "Pops"}
"Pops" = vector of population names.
"PopNum" = dataframe of population names and counts.
"Inds" = vector of sample IDs.
"Loci" = vector of Loci.
"Allele" = vector of allele values.}

\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single commma delimited row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked to the locus data by " ,  " (space space,space) and is read in as
as a single row (character).}
}
\description{
Function returns population derived allele frequencies.

Function returns Genepop file meta-data.
}

