% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_flatten.R
\name{genepop_flatten}
\alias{genepop_flatten}
\title{Flatten Genepop to dataframe}
\usage{
genepop_flatten(GenePop)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab separation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single comma delimited row of loci names followed by the locus data. Populations are
separated by "Pop". Each individual ID is linked to the locus data by " ,  " (space,space space) and is read in as
as a single row (character).}
}
\description{
Convert and return Genepop as flattend dataframe.
}

