% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_flatten.R
\name{genepop_flatten}
\alias{genepop_flatten}
\title{Flatten Genepop to dataframe}
\usage{
genepop_flatten(GenePop)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will the standard genepop format with a the first n+1 rows corresponding the the n loci names,
or a single commma deliminated row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character).}
}
\description{
Convert and return Genepop as flattend dataframe.
}

