% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_fstat.R
\name{genepop_fstat}
\alias{genepop_fstat}
\title{Convert Genepop to FSTAT format.}
\usage{
genepop_fstat(genepop, path = NULL, addworkspace = FALSE)
}
\arguments{
\item{genepop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab separation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single comma delimited row of loci names followed by the locus data. Populations are
separated by "Pop". Each individual ID is linked to the locus data by " ,  " (space,space space) and is read in as
as a single row (character).}

\item{path}{the filepath and filename of output.}

\item{addworkspace}{logical statement defining whether the converted data
should be saved as a file specified in the path (default) argument or whether it should be returned to the workspace
if returned to the workspace the object will be called "Output_fstat".}
}
\description{
Function to convert Genepop to FSTAT
}
