% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_newhybrids.R
\name{genepop_newhybrids}
\alias{genepop_newhybrids}
\title{Convert Genepop to New Hybrids format.}
\usage{
genepop_newhybrids(GenePop, path = NULL)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single commma delimitedd row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked to the locus data by "   , " (space space,space) and is read in as
as single row (character).}

\item{path}{the filepath and filename of output.}
}
\description{
Function to convert Genepop to NewHybrids
}

