% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_toploci.R
\name{genepop_toploci}
\alias{genepop_toploci}
\title{Creates a panel of the top n unlinked loci, and exports the list of loci}
\usage{
genepop_toploci(genepop, where.plink, where.pgdspider, r2.threshold = 0.2,
  fst.threshold = 0.05, ld.window = NULL, ldpop = "All",
  allocate.pgd.ram = 1, return.workspace = TRUE, save.output = FALSE)
}
\arguments{
\item{genepop}{A file path to the genepop format file you wish to create your panel from}

\item{where.plink}{A file path to the PLINK installation folder.}

\item{where.pgdspider}{A file path to the PGDspider installation folder.}

\item{r2.threshold}{The minimum r^2 threshold to consider a pair of loci to be in LD}

\item{fst.threshold}{The minimum FST threshold required to retain a locus}

\item{ld.window}{Number of adjacent SNPs to compare each SNP against for LD - default is NULL, which translates to a window size of 99999, which essentially asks to compare each SNP against all others}

\item{ldpop}{A string which populations (default: "All") will be used to calculate linkage disequilibrium. Names must match names returned by genepop_detective().}

\item{allocate.pgd.ram}{An integer value in GB to specify the maximum amount of RAM to allocate to PGDspider. The default is 1 GB, which should be sufficient for most analyses.}

\item{return.workspace}{(default: TRUE) Logical query to return the output to the workspace}

\item{save.output}{Logical query (default: FALSE) to save the output to the same location as the file being analyzed. Each of the outputs of the function will be saved as a separate file with the file name of the orginal data appended with "_Linkages", "Loci_FST", and "Unlinked_Loci_FST" for the pairwise linked loci along with their r^2, all loci with their global Fst, and only the top unlinked loci with their Fst respectively.}
}
\description{
Extract the genotypes of individuals at the top n (by Fst) unlinked loci. The default threshold of r2>0.2 is employed for assigning 'linked' loci (default for plink).
}

