% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_toploci.R
\name{genepop_toploci}
\alias{genepop_toploci}
\title{Creates a panel of the top n unlinked loci, and exports the list of loci}
\usage{
genepop_toploci(GenePop, LDpop = "All", panel.size = NULL,
  r2.threshold = 0.2, ld.window = NULL, where.PLINK, where.PGDspider,
  allocate.PGD.RAM = 1)
}
\arguments{
\item{GenePop}{A file path to the GENEPOP format file you wish to create your panel from}

\item{LDpop}{A string which populations (default: "All") will be used to calculate linkage disequilibrium. Names must match names returned by genepop_detective().}

\item{panel.size}{An integer number of loci to include in the panel. If not specified all loci will be returned,}

\item{r2.threshold}{The minimum r^2 threshold to consider a pair of loci to be in LD}

\item{ld.window}{Number of adjacent SNPs to compare each SNP against for LD - default is NULL, which translates to a window size of 99999, which essentially asks to compare each SNP against all others}

\item{where.PLINK}{A file path to the PLINK installation folder.}

\item{where.PGDspider}{A file path to the PGDspider installation folder.}

\item{allocate.PGD.RAM}{An integer value in GB to specify the maximum amount of RAM to allocate to PGDspider. The default is 1 GB, which should be sufficient for most analyses.}
}
\description{
Extract the genotypes of individuals at the top n (by Fst) unlinked loci. The default threshold of r2>0.2 is employed for assigning 'linked' loci (default for plink).
}

