% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepop_treemix.R
\name{genepop_treemix}
\alias{genepop_treemix}
\title{Convert a Genepop to input required for phython processing to TREEMIX.}
\usage{
genepop_treemix(GenePop, where.PGDspider, where.PLINK, allocate.PGD.RAM = 1,
  keep_inter = FALSE, path)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab separation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard Genepop format with the first n+1 rows corresponding to the n loci names,
or a single comma delimited row of loci names followed by the locus data. Populations are
separated by "Pop". Each individual ID is linked to the locus data by " ,  " (space,space space) and is read in as
as a single row (character).}

\item{where.PGDspider}{A file path to the PGDspider installation folder.}

\item{where.PLINK}{A file path to the PLINK installation folder.}

\item{allocate.PGD.RAM}{An integer value in GB to specify the maximum amount of RAM to allocate to PGDspider. The default is 1 GB, which should be sufficient for most analyses.}

\item{keep_inter}{A logical condition statement (default : FALSE) specifying whether to keep the map, ped, and clustering files generated during the conversion.}

\item{path}{file path to directory where the gzipped Treemix input file will be saved.}
}
\description{
Convert Genepop to within-clusters binary PED file for TREEMIX.
}

