% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenePopData}
\alias{GenePopData}
\title{Documentation for the dataset used inside genepopedit
This dataset contains genomic data coded as 3 digit alleles.}
\format{A data frame with 411 rows and 1 variable: representing a genepop file with
100 Single Nucleotide Polymorphisms (SNPs).}
\source{
bla bla bla.
}
\usage{
GenePopData
}
\description{
Documentation for the dataset used inside genepopedit
This dataset contains genomic data coded as 3 digit alleles.
}
\details{
\itemize{
\item Markers: 100 SNP
\item Populations: 10
\item Samples: 30 indivuals/population
}
}
\keyword{datasets}

