% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgdSpideR.R
\name{pgdSpideR}
\alias{pgdSpideR}
\title{Execute pgdSpider conversions in R}
\usage{
pgdSpideR(input, input_format, output, spid, output_format, where.PGDspider)
}
\arguments{
\item{input}{the full path to the input file.}

\item{output}{the full file path to the desired output. (e.g. GENEPOP or PED) these must match the dropdown menus in pgdSpider.}

\item{spid}{the fill file path to the spid file generated in pgdSpider.}

\item{where.PGDspider}{A file path to the PGDspider installation folder.}

\item{input_type}{what are you converting from (e.g. GENEPOP or PED) these must match the dropdown menus in pgdSpider.}

\item{output_type}{what are you converting to (e.g. GENEPOP or PED). This must match the dropdown menus in pgdSpider.}
}
\description{
Function to between file types in R.
}

