% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genepop.R
\name{subset_genepop}
\alias{subset_genepop}
\title{Genepop edit}
\usage{
subset_genepop(GenePop, subs = NULL, keep = TRUE, dirname, sPop = NULL)
}
\arguments{
\item{GenePop}{the genepop file to be manipulated. This will the standard
genepop format with a the first n+1 rows corresponding the the n loci names
followed by the locus data. Populations are seperated by "Pop".
Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character)
e.g.
Stacks Ver 1.0
1
2
3
Pop
BON01  , 120120 110110 110110
BON02  , 100100 110110 110110
Pop
TAG01  , 120120 110110 110110
...}

\item{subs}{he loci names of interest or a vector which corresponds the the order of which
they appear in the genepop file.
These can be either the order by which they occur or the exact name of the loci
e.g. subs <- c(1,2,3,4) would return the first 4 loci &
subs <- c("190-56","145_21",456_12") would return loci with defined names.}

\item{keep}{logical vector which defines whether you want to remove the loci or keep them.
the default is to keep them (keep <- TRUE) assuming you are removing neutral markers
and only keeping the subs}

\item{dirname}{directory where the output file will be saved.}

\item{sPop}{is the populations of interest. Note these are specified in the order which they appear in the
original Genepop file. i.e. first pop = 1 second pop = 2  or the text based origin
Examples: numeric - \code{sPop <- c(1,3,4,7)}
          text- \code{sPop <- c("BMR", "GRR","GHR","TRS")}.}
}
\description{
Function for the manipulation of genopop format SNP datasets
}

