% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genepop.R
\name{subset_genepop}
\alias{subset_genepop}
\title{Genepop subset loci and populations}
\usage{
subset_genepop(GenePop, subs = NULL, keep = TRUE, sPop = NULL, path)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will be the standard genepop format with the first n+1 rows corresponding to the n loci names,
or a single commma delimited row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked to the locus data by "   , " (space space,space) and is read in as
as a single row (character).}

\item{subs}{he loci names of interest or a vector which corresponds the the order of which
they appear in the genepop file.
These can be either the order by which they occur or the exact name of the loci
e.g. subs <-c(1,2,3,4) would return the first 4 loci
& subs <- c("190-56","145_21",456_12") would return loci with defined names.}

\item{keep}{logical vector which defines whether you want to remove the loci or keep them.
The default is to keep them keep <- TRUE assuming you are removing neutral markers
and only keeping the subs.}

\item{sPop}{is the populations of interest. Note these are specified in the order which they appear in the
original Genepop file. i.e. first pop = 1 second pop = 2
Examples: numeric - sPop <- c(1,3,4,7) or
the population ID (alpha-numeric code before the underscore). Here we assume conventional
naming of "Population_sample#" e.g. (Aqua01_05: population Aqua01 & sample #5).
          text- sPop <- c("Aqua01", "GRR","GHR","TRS").}

\item{path}{the filepath and filename of output.}
}
\description{
Function to subset loci and populations
}

