% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genepop_aggregate.R
\name{subset_genepop_aggregate}
\alias{subset_genepop_aggregate}
\title{Genepop subset, combine, and reorder populations}
\usage{
subset_genepop_aggregate(GenePop, dirname, subs = NULL, keep = TRUE,
  agPopFrame)
}
\arguments{
\item{GenePop}{the genepop file to be manipulated. This will the standard
genepop format with a the first n+1 rows corresponding the the n loci names
followed by the locus data. Populations are seperated by "Pop".
Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character)
e.g.
Stacks Ver 1.0
1
2
3
Pop
Pop01_01  , 120120 110110 110110
Pop01_02  , 100100 110110 110110
Pop
Pop02_01  , 120120 110110 110110
...}

\item{subs}{he loci names of interest or a vector
subs <- c("190-56","145_21",456_12") would return loci with these defined names.}

\item{keep}{logical vector which defines whether you want to remove the loci or keep them.
the default is to keep them keep <- TRUE assuming you are removing neutral markers
and only keeping the subs}

\item{agPopFrame}{a dataframe or path to a csv.
This dataframe contains two columns: Column 1 corresponds to the population names.
Here we consider the alpha-numeric characters before the first underscore '_' to be the population name.
so that IDs are "Population_sample#" (e.g. Aqua23_04 = Population Aqua23, individual 4).
These names can be obtained using the genepop_detective function.
The next column has grouping variables. If you don't want to change the grouping just repeat original name.
If the input is a dataframe object from the workspace it must be a data.frame object and therefore will have headers.
e.g.
Opop   AgPop
AAA    Pop1
BBB    Pop1
CCC    CCC
DDD    Pop2
EEE    EEE
FFF    Pop2
GGG    GGG
AAA/BBB & DDD/FFF would be clustered together between population flags in genepop.}

\item{path}{to the output .txt file (e.g."c:/Users/Yourname/Documents/output.txt")}
}
\description{
Function to cluster populations together and remove specific loci
}

