% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genepop_aggregate.R
\name{subset_genepop_aggregate}
\alias{subset_genepop_aggregate}
\title{Genepop subset and combine/rename populations}
\usage{
subset_genepop_aggregate(GenePop, subs = NULL, keep = TRUE, dirname,
  agPopFrame)
}
\arguments{
\item{GenePop}{the genepop file to be manipulated. This will the standard
genepop format with a the first n+1 rows corresponding the the n loci names
followed by the locus data. Populations are seperated by "Pop".
Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character)
e.g.
Stacks Ver 1.0
1
2
3
Pop
Pop01_01  , 120120 110110 110110
Pop01_02  , 100100 110110 110110
Pop
Pop02_01  , 120120 110110 110110
...}

\item{subs}{he loci names of interest or a vector which corresponds the the order of which
they appear in the genepop file.
These can be either the order by which they occur or the exact name of the loci
e.g. subs <- c(1,2,3,4) would return the first 4 loci &
subs <- c("190-56","145_21",456_12") would return loci with defined names.}

\item{keep}{logical vector which defines whether you want to remove the loci or keep them.
the default is to keep them (keep <- TRUE) assuming you are removing neutral markers
and only keeping the subs}

\item{dirname}{directory where the output file will be saved.}

\item{agPopFrame}{a dataframe or path to a csv. This must be specified in this function
this dataframe contains two columns. Column 1 corresponds to the population names. These names
should match the individual IDs (e.g. BON01  , 110110 120120 -- would be 'BON'). The next column
has the new names that you want to replace. If you don't want to change the name then just repeat
from column one in that row. **Note if this is a reference to a path then there should be column headers
though the headers do not need to match this example. If the input is a dataframe object from the
workspace it must be a data.frame object and therefore will have headers.
e.g.
Opop   AgPop
BON    BON
BRA    BON
EDN    EDN
CRA    CRA
MAL    BON
TRY    CRA}
}
\description{
Function for the manipulation of genopop format SNP datasets and renaming of populations
}

