% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genepop_individuals.R
\name{subset_genepop_indiv}
\alias{subset_genepop_indiv}
\title{Genepop remove or keep specific sample IDs}
\usage{
subset_genepop_indiv(GenePop, dirname, indiv = NULL, keep = FALSE)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will the standard genepop format with a the first n+1 rows corresponding the the n loci names,
or a single commma deliminated row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character).}

\item{dirname}{path to the output .txt file (e.g."c:/Users/Yourname/Documents/output.txt")}

\item{indiv}{names of individual samples
These can be either the order by which they occur or the exact name of the loci
indiv <- \code{c("Pop01_01","Pop03_15","Pop16_02")} would individuals with these sample names.}

\item{keep}{logical vector which defines whether you want to remove the loci or keep them.
the default is to remove them (keep <- FALSE)}
}
\description{
Function for the manipulation of genopop format SNP datasets
}

