% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_genepop_rename.R
\name{subset_genepop_rename}
\alias{subset_genepop_rename}
\title{Genepop subset and rename populations}
\usage{
subset_genepop_rename(GenePop, nameframe, renumber = FALSE, path)
}
\arguments{
\item{GenePop}{the genepop data to be manipulated. This can be either a file path
or a dataframe read in with tab seperation, header=FALSE , quote="", and stringsAsFactors=FALSE.
This will the standard genepop format with a the first n+1 rows corresponding the the n loci names,
or a single commma deliminated row of loci names followed by the locus data. Populations are
seperated by "Pop". Each individual ID is linked the the locus data by "  , " and is read in as
as single row (character).}

\item{nameframe}{a dataframe or path to a csv. This must be specified in this function
this dataframe contains two columns. Column 1 corresponds to the population names. These names
should match the individual IDs (e.g. BON01  , 110110 120120 -- would be 'BON'). The next column
has the new names that you want to change or rename. If you don't want to change the name then just repeat
from column one in that row. **Note if this is a reference to a path then there should be column headers
though the headers do not need to match this example. If the input is a dataframe object from the
workspace it must be a data.frame object and therefore will have headers.
This function assumes that each population will have a unique name. Names in this case are
comprised of alpha characters and not numbers.
(e.g. Pop01_01 and Pop02_01 would each be considered 'Pop' for the population name)
e.g.
Opop   Rename
BON    BON
BRA    BON
EDN    EDN
CRA    CRA
MAL    BON
TRY    CRA}

\item{renumber}{is a logical (default=FALSE) defining whether you want to change the sample unique identitiy
i.e. sample number - BON_01 where 01 is the unique qauntity. If multiple populations are combined this will
prevent two samples from having the same name.}

\item{path}{the filepath and filename of output.}
}
\description{
Function for the manipulation of genopop format SNP datasets and renaming of populations
}

